Admin_Auth/requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.5.0
django-cors-headers>=4.3.1 